[main index] check REQUEST eurekadb variable
To change the db into the webservice, the REQUEST variable is not available,
and SESSION must be use.
Then the main index file checks if REQUEST for eurekadb is present to create
SESSION entry. In the reverse case, the main index file removes the SESSION
entry.

This modification allows to change the db by the url.

SECURITY tests must be done on this feature.